Initial pass of the py xapp frame
This is not yet completed:
- Docs will come soon
- More unit test coverage will come soon
- More functionality will come soon

However, to avoid a gargantuan commit at the beginning that solves world
hunger, we cut off here on a basically usable framework.

NOTE: a significant portion of this code is *already reviewed* as it
came out of A1. For example, the entire SDLWrapper class. This code will
be deleted from A1 when this is in pypi.

Issue-ID: RIC-228
Change-Id: I9f87b0178cf5fb1cb6b988c29ce11218ed575607
Signed-off-by: Tommy Carpenter <tc677g@att.com>