Update patch set 1
Patch Set 1:

it's because of the triggers in the CICD job

basically, this didnt make a code change and didnt bump the version. so the attempted push to pypi, which gets attempted on merge, fails because the version already exists. But, it's fine, because the code didnt change and so this doesn't need to go to pypi

The fix for this is to carefully configure the triggers in the CICD yaml for this, which I haven't done yet. It's a fight of correctness vs laziness since we can just ignore the failure

your curiosity is a shining beacon in my life guy, never forget

for real fun, you can add Chris Lott onto this and he will say I TOLD YOU SO to me, here:

Patch-set: 1
CC: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>