Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> So it's probably not timeouts. Could it be something with the loopback interface, 127.0.0.1, in this Openstack container?  Do any of the other tests try to send to that IP as a destination?

It's a possibility.  RTS is setup by using the source address that is placed into the origiginal message. By default this is the hostname:port and if that isn't resolving then rts messages won't flow back.   That was the thing I wanted to check on with the RMR tests; there is a way to force that value because in the k8 world hostname isn't routable and the service name has to be forced in.  For the RMR vetting this isn't needed, but I don't know if the builder/tester here is a different flavour.

The other possible issue is whether or not these are being driven from the same process or not.  That shouldn't matter, but that would be a difference between the RMR tests -- RMR starts two processes for it's RTS testing.

Patch-set: 8
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>