Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> I see a comment that says "localhost does not work in jenkins" (meaning, in a build minion).  That suggests to me, using the IP 127.0.0.1 will not work either.
> 
> How about calling hostname() and using the value returned by it, just for kicks $h!t$ and giggles?  Is that even dumber?


If it's possible to set this environment variable on both processes should force a known source into the message that goes out:

RMR_SRC_ID=127.0.0.1

I assume this is possible since SEED_RT should be supplied too.

I'm pretty sure that the issue is that the source name that is available to the sender (and this is what is returned by the system to RMR itself) cannot be resolved when RTS is used. 

Another clue would be to check the status that is returned in the message returned by RTS;



Meeting is winding down, so I can test locally in a few minutes.

Patch-set: 8
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>