Update patch set 11
Patch Set 11:

Some days I hate gerrit -- I can't find your message to reply to...

After the RTS the source will indeed be the source of the sender, not the source of the current process.  This allows you to use that message to call RTS multiple times.  So, seeing that the source after RTS is that of the other process is a good thing.

Patch-set: 11
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>