Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> scott, the test sender does indeed send a 0 length payload, did you say something about that?

Yes.  I can't remember if that's valid from RMR's point of view, but we know that message is getting in because we see the RTS work.

I just ran on u1804 with py 3.8:

________________________________________________________________________________ summary ________________________________________________________________________________
  code: commands succeeded
  flake8: commands succeeded
  docs: commands succeeded
  docs-linkcheck: commands succeeded
  congratulations :)


I think this is a jenkins environment issue, but I don't have a clue.  Are we sure jenkins is 3.8 py?

I need to walk away for a bit. My eyes are giving me fits at the moment.

Patch-set: 11
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>