Move rmr python here.
The decision was made to move rmr python into the xapp frame.
It's module name has not been changed otherwise, so apps transitioning
to this should only need to prefix "rmr" with "ricxappframe." to
transition (and, replace rmr with ricxappframe if they are not already
using the framework).

Some small changes have been made in the unit tests, such as port
numbers and monkeypatching statements, but otherwise all "new" code is
identical to rmr python.

Issue-ID: RIC-228
Change-Id: I5e80cfd39c1511db93b95ac5e442a2acfc0733a2
Signed-off-by: Tommy Carpenter <tc677g@att.com>