Update patch set 2
Patch Set 2:

one of the consequences of this move is a bit of version number confusion.  The xapp python framework is sitting pre-1.0; here it is changed to 0.8.0.  The RMR python binding was flying along at 4.0.0, looking kind of mature and stable.  You all feel this is just an acceptable cost?

Patch-set: 2
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>