Update patch set 3
Patch Set 3:

I am guessing that as soon as this move-change is merged, the old rmr-python bindings will be declared to be deprecated, and all users must switch when they want the next version.  Is that a fair statement?

I would be glad to add the Nokia xapp-python team members here, just to get them ready for the upcoming change, but I have not been corresponding with them and don't know whom to add.

Patch-set: 3
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>