Update patch set 1
Patch Set 1:

that would be a scott daniels questions..

there is a pydoc autogeneration process in rmr python that generates documentation of the complete API. If you look in the rmr repo, there is a Dockerfile called Doc-Gen that does this. I don't know how Scott integrated that.

This is the piece of work I mentioned yesterday; the docs have to get moved over into this repo, somehow. All the autodoc stuff is still in the code, so perhaps we keep it and just generate a static .RST file. 

this is on the list of things to do, however cutting over A1 is ahead on that list to make sure this works. 

then move the docs

then delete the rmr code

Patch-set: 1