Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Chris, when you are the master here, fixing that merge job is probably the first thing you will do =)
> > 
> > basically, merging the "release" file should not try to trigger a pypi push, it's a trigger problem
> 
> This comment just baffles me.  Knowing how much you hate the push-on-merge behavior, I offered that fix last week, but you added -1 to the change here and it remains blocked:
> 
> https://gerrit.o-ran-sc.org/r/c/ci-management/+/3095

I don't hate push on merge when I want push on merge lol

Basically, *every other merge* works properly. So, pick your poision. If that is merged, it will fix this merge but require extra work on the other merges.

But OK, yes you can just merge that and do the extra work on the other merge.

Patch-set: 1