Update patch set 3
Patch Set 3:

A note about why this does not work without rmr installed: autodoc scans the code to build docs. Scanning the code will fail if RMR is not installed since it will never make it past the import statement.

Locally, a developer can build the docs which will produce a very pretty html of the rmr API. But alas this requires rmr.

Patch-set: 3