Update patch set 3
Patch Set 3:

Chris Lott you have the final word here, I understand you aren't thrilled with this, but it never worked in rmr either.

If this is merged, I will submit a PR to rmr deleting all the code.

If this is not merged, there is still stuff in rmr, like the examples, that are _not_ in here.

I recommend merging this despite the API doc not working, *because* it also does not work in rmr therefore there is no "loss" or "downgrade" here.

If perfect is the enemy of .."ok" here, then you can deal with this how you'd like in the future.

Patch-set: 3