Update patch set 4
Patch Set 4:

This is now more and less technically elegant than before.

More, because the API documentation is generated and published automatically, guaranteeing that the ReadTheDocs pages exactly matche the code.

Less, because the header-file constants that RMR exposes via a method call are duplicated in the python code.  But those values are frozen in the C world, so they won't ever change in the Python world either.

Patch-set: 4