Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> This is a pick-your-poison kind of issue.  
> 
> Today the RMR package publishes constants by querying the library at load (import) time.  When the autodoc generator runs, it imports the packages.  That triggers the constant-query method.  That fails in ReadTheDocs. Bottom line seems to be, if I want good automatically generated doc, the query at import time must not happen.
> 
> There are various ways to stop the query at import time.  Tommy has rejected so far:
> 
> 1) a cached copy of RST API documentation
> 2) changing the RMR file to have fixed constants
> 3) breaking the RMR API - I was thinking to add a method getConstant(key) that user code will query

Oh, so this is a problem because the doc generator doesn't just parse the files looking for doc to format and spit out, but actually tries to  load the code?   Man, that just seems backwards to me.

If this is a doc thing, why are the values important?  Shouldn't the doc just have the constant names which should not change unless there is a very very good reason.   Is there a way to do that and keep things happy?

Patch-set: 5
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>