Update patch set 5
Patch Set 5:

Since Scott and others are passionate about not duplicating header-file values in Python, then it seems to come down to the method for client code to fetch code constants.  Today these are Python constants,  values populated during package import time.  Since there are relatively few constants, and relatively few users, I think it would be a tolerable level of breakage to add a getConstant(key) method and drop the code constants, as part of publishing a version that's compatible with 3.7.1 where there is no max message size.

Patch-set: 5