Update patch set 5
Patch Set 5:

Then what I need here is a Python trick.  I can preserve the existing package API and I can keep
the autodoc generator working IF the rmr.py module can discover when the autodoc generator is active. 
In other words, I'd like the code to detect when the rmrclib object is actually a mock (not a
real object), and in that situation, don't call the real getConstants() method.

Patch-set: 5