Update patch set 5
Patch Set 5:

Thanks to Matti and Scott, RMR man pages are now available in ReadTheDocs and I added a link to that page.  I plan no further changes here, please review.

I changed the formatting of the release notes because the previous markup forced the English text to be processed as code. One consequence was highlighting of common reserved words like "for". The browser also displayed scrollable boxes for every release to handle the long lines, because Sphinx doesn't re-flow/wrap the lines. The release notes are now much easier to read in a browser.

Patch-set: 5