Use blocking get call w/ timeout to read msg queue
Add parameters to the queue get method in the xapp_frame loop that reads
messages so it waits for a message to arrive and occasionally checks for
the end-loop flag, instead of spinning the CPU at 100% while waiting.

Upgrade all to use the latest RMR, version 4.0.5.

Tweak the example xapps to emit their names in log messages.

Improve documentation especially the package overview shown at PyPI.

Issue-ID: RIC-354
Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>
Change-Id: I08692e6ef60d199cb0b92c1c99740ae808b8885c