Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I think the verification job failure is related to an LF script (release-job.sh) hard setting the version of idna but the min version required by  whatever "lftools" are is higher.   I'll open a ticket with LF given that they'll have to change their stuff.
> 
> I'm sorry about late response of mine, I just forgot the check this review earlier...
> There is also this error log:
> 08:43:34 INFO: Fetching console log from https://logs.o-ran-sc.org/production/vex-sjc-oran-jenkins-prod-1/ric-plt-xapp-frame-py-pypi-stage-master/8
> 08:43:34 Build step 'Execute shell' marked build as failure
> 
> I think currently we do not have staging log #8 available:
> https://jenkins.o-ran-sc.org/view/ric-plt-xapp-frame-py/job/ric-plt-xapp-frame-py-pypi-stage-master/

I saw that too, and am a bit confused.  The way I read it, the previous release used log #7, and I would have expected that there would have been changes to the code since then that generated the next log, but that seems not to have happened; the last log for that job in Jenkins is showing #7;  it seems incorrect to use the same log as used before, but I don't exactly know what information the release job is looking for in the old merge logs.

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>