Fix ping-pong demo (bump to RMR 4.8.3, typo)
This commit relies on https://gerrit.o-ran-sc.org/r/c/it/dev/+/8495
to be merged. The Python xApp framework seems to have evolved to a
point where the original RMR version in the ping pong example is
incompatible with the framework (or even the most recent available
builder image, hence the other issue referenced above). With the
new builder image tag in place, there is still a bug in pong_xapp
that prevents the example from executing, which is fixed in this
commit.

Change-Id: I6579d85bf73eb579c51a7fb21fe621ad021a858b
Signed-off-by: Johannes K Becker <johannes@caur.us>